+++ src/test-integration/kotlin/book/ListBookIntegrationTest.kt
@@ -79,7 +79,7 @@ class ListBookIntegrationTest(
         title = "Book One",
         summary = "First book summary",
         tableOfContents = "TOC 1",
-        price = 300,
+        price = 300.toBigDecimal(),
         numberOfPages = 200,
         isbn = "111-111-111",
         publishDate = LocalDate.now().plusDays(15),
