+++ src/test/kotlin/br/com/study/codedesignpractice/book/controller/request/BookRequestTest.kt
@@ -67,7 +67,7 @@ class BookRequestTest {
         title = "Effective Kotlin",
         summary = "A book about effective Kotlin practices.",
         tableOfContents = "1. Introduction\n2. Basics",
-        price = 30,
+        price = 30.toBigDecimal(),
         numberOfPages = 200,
         isbn = "123-4567890123",
         publishDate = LocalDate.now().plusDays(1),
