+++ src/test/kotlin/br/com/study/codedesignpractice/book/repository/BookRepositoryTest.kt
@@ -40,7 +40,7 @@ class BookRepositoryTest @Autowired constructor(
             title = "Book Title",
             summary = "Book summary",
             tableOfContents = "Markdown table of contents",
-            price = 250,
+            price = 250.toBigDecimal(),
             numberOfPages = 150,
             isbn = "123-456-789",
             publishDate = LocalDate.now().plusDays(10),
